Se crea la pagina de Ubicación (Geolocalización) y se configura como página de aterrizaje para el login.
