+++ Aplicacion/FrontSeguBici/segubicifront/src/pages/home/home.ts
@@ -1,6 +1,8 @@
 import { Component } from '@angular/core';
 import { NavController } from 'ionic-angular';
 import { RegistrarsePage } from '../registrarse/registrarse';
+import { RestProvider } from '../../providers/rest/rest';
+import { UbicacionPage } from '../ubicacion/ubicacion';
 
 
 @Component({
@@ -9,15 +11,37 @@ import { RegistrarsePage } from '../registrarse/registrarse';
 })
 export class HomePage {
 
-  constructor(public navCtrl: NavController) {
+  usuario: String;
+  clave: String;
 
+  constructor(public navCtrl: NavController, public restProvider: RestProvider) {
   }
 
   registrarse() {
 
-   //this.navCtrl.push(LoginPage) es un metodo para configurar el boton
-    this.navCtrl.setRoot(RegistrarsePage)
+    //this.navCtrl.push(LoginPage) es un metodo para configurar el boton
+    this.navCtrl.push(RegistrarsePage)
 
   }
 
+  ionViewDidLoad() {
+    if(window.localStorage["token"]){
+      this.navCtrl.setRoot(UbicacionPage);
+    }
+  }
+
+  iniciarSesion() {
+    var data = { 'username': this.usuario, 'password': this.clave };
+    this.restProvider.login(data)
+      .then((data:any) => {
+        window.localStorage['token'] = data.key; // para guardar el token en el local storage -consola-aplicacion
+        this.navCtrl.push(UbicacionPage);
+      }, (err) => {
+        console.log(err);
+      });
+  }
+  
+  mostrarUbicacion() {
+    this.navCtrl.push(UbicacionPage);
+    }
 }
