--- Aplicacion/FrontSeguBici/segubicifront/src/pages/login/login.module.ts
@@ -1,13 +1,0 @@
-import { NgModule } from '@angular/core';
-import { IonicPageModule } from 'ionic-angular';
-import { LoginPage } from './login';
-
-@NgModule({
-  declarations: [
-    LoginPage,
-  ],
-  imports: [
-    IonicPageModule.forChild(LoginPage),
-  ],
-})
-export class LoginPageModule {}
