--- Aplicacion/FrontSeguBici/segubicifront/src/pages/login/login.scss
@@ -1,3 +1,0 @@
-page-login {
-
-}
